Hechos 4,1 y 4,5

+++ src/main/java/U1/T4/ej4_1.java
@@ -4,7 +4,7 @@ public class ej4_1 {
     public static void main(String[] args) {
         short num1 = Short.MAX_VALUE;
         System.out.println("Valor Mínimo Short: " + Short.MAX_VALUE);
-        short num2 = num1 + 1;
+        short num2 = (short) (num1 + 1);
         System.out.println(num2);
     }
 }
